Cleanup and resolved requested changes
