+++ patches/tModLoader/Terraria/Item.cs.patch
@@ -233,50 +233,44 @@
  			float num14 = 1f * num2 * (2f - num4) * (2f - num7) * num5 * num3 * num6 * (1f + (float)num8 * 0.02f);
  			if (num == 62 || num == 69 || num == 73 || num == 77)
  				num14 *= 1.05f;
-@@ -1064,21 +_,36 @@
+@@ -1064,6 +_,23 @@
  			if (num == 65 || num == 72 || num == 76 || num == 80 || num == 68)
  				num14 *= 1.2f;
  
 +			if (num >= PrefixID.Count)
 +				ModPrefix.GetPrefix((byte)num)?.ModifyValue(ref num14);
 +
-+			if (rare <= 11) {
--			if ((double)num14 >= 1.2)
-+				if ((double)num14 >= 1.2)
--				rare += 2;
-+					rare += 2;
--			else if ((double)num14 >= 1.05)
-+				else if ((double)num14 >= 1.05)
--				rare++;
-+					rare++;
--			else if ((double)num14 <= 0.8)
-+				else if ((double)num14 <= 0.8)
--				rare -= 2;
-+					rare -= 2;
--			else if ((double)num14 <= 0.95)
-+				else if ((double)num14 <= 0.95)
--				rare--;
-+					rare--;
++			if (rare >= ItemRarityID.Count) {
++				int offset = 0;
++				if (num14 >= 1.2)
++					offset += 2;
++				else if (num14 >= 1.05)
++					offset++;
++				else if (num14 <= 0.8)
++					offset -= 2;
++				else if (num14 <= 0.95)
++					offset--;
++				rare = RarityLoader.GetRarity(rare).ModifyRarityOffset(offset);
++				goto SkipVanillaRarity;
 +			}
-+			else {
-+				if ((double)num14 >= 1.2)
-+					rare = ModRarity.GetRarity(rare).RarityPlusTwo();
-+				else if ((double)num14 >= 1.05)
-+					rare = ModRarity.GetRarity(rare).RarityPlusOne();
-+				else if ((double)num14 <= 0.8)
-+					rare = ModRarity.GetRarity(rare).RarityMinusTwo();
-+				else if ((double)num14 <= 0.95)
-+					rare = ModRarity.GetRarity(rare).RarityMinusOne();
-+			}
++
+ 			if ((double)num14 >= 1.2)
+ 				rare += 2;
+ 			else if ((double)num14 >= 1.05)
+@@ -1073,12 +_,13 @@
+ 			else if ((double)num14 <= 0.95)
+ 				rare--;
  
- 			if (rare > -11) {
++			SkipVanillaRarity:
+-			if (rare > -11) {
++			if (rare > ItemRarityID.Quest) {
  				if (rare < -1)
  					rare = -1;
  
 -				if (rare > 11)
 -					rare = 11;
-+				if (rare > ModRarity.RarityCount)
-+					rare = ModRarity.RarityCount - 1;
++				if (rare > RarityLoader.RarityCount)
++					rare = RarityLoader.RarityCount - 1;
  			}
  
  			num14 *= num14;
