+++ patches/tModLoader/Terraria/Main.cs.patch
@@ -1027,7 +1027,7 @@
  				baseColor = new Microsoft.Xna.Framework.Color((byte)(180f * num2), (byte)(40f * num2), (byte)(255f * num2), mouseTextColor);
  
 +			if (num > 11)
-+				baseColor = ModRarity.GetRarity(num).RarityColor() * num2;
++				baseColor = RarityLoader.GetRarity(num).RarityColor() * num2;
 +
 -			if (HoverItem.expert || num == -12)
 +			if (HoverItem.expert || num == ItemRarityID.Expert)
@@ -1134,7 +1134,7 @@
  							black = new Microsoft.Xna.Framework.Color((byte)(180f * num4), (byte)(40f * num4), (byte)(255f * num4), a);
  
 +						if (rare > 11)
-+							black = ModRarity.GetRarity(rare).RarityColor() * num4;
++							black = RarityLoader.GetRarity(rare).RarityColor() * num4;
 +
  						if (diff == 1)
  							black = new Microsoft.Xna.Framework.Color((byte)((float)(int)mcColor.R * num4), (byte)((float)(int)mcColor.G * num4), (byte)((float)(int)mcColor.B * num4), a);
