+++ patches/tModLoader/Terraria/PopupText.cs.patch
@@ -41,8 +41,8 @@
 -				else if (newItem.rare >= 11)
 +				else if (newItem.rare == 11)
  					Main.popupText[num2].color = new Color(180, 40, 255);
-+				else if (newItem.rare > 11)
-+					Main.popupText[num2].color = ModRarity.GetRarity(newItem.rare).RarityColor();
++				else if (newItem.rare >= ItemRarityID.Count)
++					Main.popupText[num2].color = RarityLoader.GetRarity(newItem.rare).RarityColor();
  
 +				Main.popupText[num2].rarity = newItem.rare;
  				Main.popupText[num2].expert = newItem.expert;
@@ -55,8 +55,8 @@
  			else if (master)
  				color = new Color(255, (byte)(Main.masterColor * 200f), 0, Main.mouseTextColor);
 +
-+			if (rarity > ItemRarityID.Purple && ModRarity.GetRarity(rarity).Animated())
-+				color = ModRarity.GetRarity(rarity).RarityColor();
++			if (rarity > ItemRarityID.Purple)
++				color = RarityLoader.GetRarity(rarity).RarityColor();
  
  			bool flag = false;
  			string text = name;
